🚀 Refactor Diffusion Models for Consistency and Elegance 🎨
🔧 Unified the structure of diffusion model classes to follow a standard pattern.
🔍 Moved model loading logic inside the model classes for enhanced encapsulation.
🔄 Switched to using the `__call__` method for initiating the generation process.
🌐 Integrated with the `diffusion_pipeline` function for seamless model instantiation.
🧹 Cleaned up repetitive code and reduced redundancy.
