--- custom_diffusion/__init__.py
@@ -1,3 +1,0 @@
-from . import pipelines, utils
-
-__version__ = "0.1.10"
