+++ diffusersplus/__init__.py
@@ -1,0 +1,4 @@
+from . import pipelines, utils
+from .automodel import diffusion_pipeline
+
+__version__ = "0.0.1"
